# perf: recycle intermediate allocations in MGet/MSet helpers