--- a/internal/cmds/builder_test.go
+++ b/internal/cmds/builder_test.go
@@ -19,6 +19,21 @@ retry:
 	}
 }
 
+func TestPutCompletedForce(t *testing.T) {
+retry:
+	cs1 := get()
+	cs1.s = append(cs1.s, "1", "1", "1", "1", "1")
+	cs1.r = 1 // pin
+	PutCompletedForce(Completed{cs: cs1})
+	cs2 := get()
+	if cs1 != cs2 {
+		goto retry
+	}
+	if len(cs2.s) != 0 {
+		t.Fatalf("PutCompletedForce doesn't clean the CommandSlice")
+	}
+}
+
 func TestPutCacheable(t *testing.T) {
 retry:
 	cs1 := get()