Using the same benchmark of https://github.com/redis/rueidis/pull/287
```
▶ benchstat old.txt new.txt
goos: darwin
goarch: arm64
pkg: rueidis-benchmark/sep
                                    │   old.txt   │              new.txt               │
                                    │   sec/op    │   sec/op     vs base               │
/OneNode/128_RueidisDoMulti-10        41.34µ ± 0%   41.49µ ± 0%  +0.36% (p=0.005 n=10)
/OneNode/128_RueidisMGet-10           21.53µ ± 0%   20.86µ ± 0%  -3.10% (p=0.000 n=10)
/OneNode/128_RueidisMGetCache-10      10.81µ ± 1%   10.19µ ± 1%  -5.75% (p=0.000 n=10)
/OneNode/128_RueidisDoMultiCache-10   4.097µ ± 1%   4.078µ ± 1%       ~ (p=0.137 n=10)
/Cluster/128_RueidisDoMulti-10        30.50µ ± 4%   29.52µ ± 3%       ~ (p=0.075 n=10)
/Cluster/128_RueidisMGet-10           34.63µ ± 4%   33.94µ ± 1%       ~ (p=0.123 n=10)
/Cluster/128_RueidisMGetCache-10      9.207µ ± 2%   9.209µ ± 1%       ~ (p=0.971 n=10)
/Cluster/128_RueidisDoMultiCache-10   4.291µ ± 1%   4.363µ ± 1%  +1.68% (p=0.000 n=10)
geomean                               14.19µ        13.97µ       -1.58%

                                    │   old.txt    │               new.txt                │
                                    │     B/op     │     B/op      vs base                │
/OneNode/128_RueidisDoMulti-10        18.17Ki ± 0%   18.17Ki ± 0%        ~ (p=0.974 n=10)
/OneNode/128_RueidisMGet-10           40.23Ki ± 0%   40.23Ki ± 0%   -0.01% (p=0.003 n=10)
/OneNode/128_RueidisMGetCache-10      30.29Ki ± 1%   26.62Ki ± 0%  -12.11% (p=0.000 n=10)
/OneNode/128_RueidisDoMultiCache-10   11.20Ki ± 1%   11.02Ki ± 0%   -1.54% (p=0.000 n=10)
/Cluster/128_RueidisDoMulti-10        18.65Ki ± 0%   18.58Ki ± 0%        ~ (p=0.075 n=10)
/Cluster/128_RueidisMGet-10           35.10Ki ± 0%   33.16Ki ± 0%   -5.52% (p=0.000 n=10)
/Cluster/128_RueidisMGetCache-10      29.53Ki ± 1%   26.34Ki ± 1%  -10.82% (p=0.000 n=10)
/Cluster/128_RueidisDoMultiCache-10   11.24Ki ± 1%   10.95Ki ± 1%   -2.61% (p=0.000 n=10)
geomean                               21.95Ki        21.02Ki        -4.24%

                                    │   old.txt    │               new.txt                │
                                    │  allocs/op   │ allocs/op   vs base                  │
/OneNode/128_RueidisDoMulti-10         130.0 ±  0%   130.0 ± 0%        ~ (p=1.000 n=10) ¹
/OneNode/128_RueidisMGet-10            132.0 ±  0%   132.0 ± 0%        ~ (p=1.000 n=10) ¹
/OneNode/128_RueidisMGetCache-10       14.00 ±  0%   12.00 ± 0%  -14.29% (p=0.000 n=10)
/OneNode/128_RueidisDoMultiCache-10   10.000 ± 10%   9.000 ± 0%  -10.00% (p=0.005 n=10)
/Cluster/128_RueidisDoMulti-10         135.0 ±  0%   135.0 ± 0%        ~ (p=1.000 n=10) ¹
/Cluster/128_RueidisMGet-10            138.0 ±  0%   137.0 ± 0%   -0.72% (p=0.000 n=10)
/Cluster/128_RueidisMGetCache-10       14.00 ±  0%   13.00 ± 0%   -7.14% (p=0.000 n=10)
/Cluster/128_RueidisDoMultiCache-10    10.00 ±  0%   10.00 ± 0%        ~ (p=1.000 n=10) ¹
geomean                                39.78         38.12        -4.17%
¹ all samples are equal
```